Tests/LibWeb/Text/expected/css/constructed-style-sheets.txt

Disabled constructed style sheet applies to document: false
Constructed style sheet with media rules applies to document: true
